api/requirements.txt
pillow
requests
azure-functions
azure-storage-blob
azure-cognitiveservices-vision-customvision